packages update and int64 IDd

+++ Moonlimit.DroneAPI/Moonlimit.DroneAPI.Entity/OwnedEntity.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace Moonlimit.DroneAPI.Entity
 {
     public class OwnedEntity : BaseEntity
     {
-        public int UserId { get; set; }
-        public int CompanyAccountId { get; set; }
+        public Int64 UserId { get; set; }
+        public Int64 CompanyAccountId { get; set; }
     }
 }
